# backend/legacy-api/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for KEC Placement Portal",
        "title": "Placement Portal KEC API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "172.16.1.208",
    "basePath": "/api",
    "paths": {
...
        "/v1/admin/analytics/dashboard": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Calculates placement statistics across multiple dimensions (year, dept, gender, category, offer-type)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Analytics"
                ],
                "summary": "Get dashboard placement analytics",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by timeframe: all_time, weekly, monthly, quarterly, half_yearly, annual",
                        "name": "timeframe",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.DashboardAnalytics"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.CategoryWiseAnalytics": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "placed_students": {
                    "type": "integer"
                }
            }
        },
...
        "models.DashboardAnalytics": {
            "type": "object",
            "properties": {
                "category_wise": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.CategoryWiseAnalytics"
                    }
                },
                "department_wise": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.DepartmentWiseAnalytics"
                    }
                },
                "gender_wise": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.GenderWiseAnalytics"
                    }
                },
                "offer_type_wise": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.OfferTypeAnalytics"
                    }
                },
                "salary_wise": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.SalaryBracketAnalytics"
                    }
                },
                "top_recruiters": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.TopRecruitersAnalytics"
                    }
                },
                "year_wise": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.YearWiseAnalytics"
                    }
                }
            }
        },
...
        "models.DepartmentWiseAnalytics": {
            "type": "object",
            "properties": {
                "department_code": {
                    "type": "string"
                },
                "interested_students": {
                    "type": "integer"
                },
                "placed_students": {
                    "type": "integer"
                },
                "placement_percentage": {
                    "type": "number"
                },
                "total_students": {
                    "type": "integer"
                }
            }
        },
...
        "models.GenderWiseAnalytics": {
            "type": "object",
            "properties": {
                "gender": {
                    "type": "string"
                },
                "placed_students": {
                    "type": "integer"
                }
            }
        },
...
        "models.OfferTypeAnalytics": {
            "type": "object",
            "properties": {
                "offer_type": {
                    "type": "string"
                },
                "placed_students": {
                    "type": "integer"
                }
            }
        },
...
        "models.SalaryBracketAnalytics": {
            "type": "object",
            "properties": {
                "bracket": {
                    "type": "string"
                },
                "placed_students": {
                    "type": "integer"
                }
            }
        },
...
        "models.TopRecruitersAnalytics": {
            "type": "object",
            "properties": {
                "company_name": {
                    "type": "string"
                },
                "placed_students": {
                    "type": "integer"
                }
            }
        },
...
        "models.YearWiseAnalytics": {
            "type": "object",
            "properties": {
                "batch_year": {
                    "type": "integer"
                },
                "interested_students": {
                    "type": "integer"
                },
                "placed_students": {
                    "type": "integer"
                },
                "placement_percentage": {
                    "type": "number"
                },
                "total_offers": {
                    "type": "integer"
                },
                "total_students": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}